feat: Refactor SymbolCard for granular component structure
