+++ frontend/src/components/SymbolCard/SymbolCard.tsx
@@ -1,10 +1,11 @@
 import './symbolCard.css';
 import { useAppSelector } from '@/hooks/redux';
 import { selectShowCardInfo } from '@/store/dashboardOptionsSlice';
-import { useState, useEffect } from 'react';
 import SymbolCardHeader from '@/components/SymbolCard/src/SymbolCardHeader';
 import PriceDisplay from '@/components/SymbolCard/src/PriceDisplay';
 import SymbolCardInfo from '@/components/SymbolCard/src/SymbolCardInfo';
+import { getCardClass } from '@/utils/get-card-class';
+import { useCardShake } from '@/hooks/use-card-shake';
 
 type SymbolCardProps = {
   id: string;
@@ -19,38 +20,11 @@ const SymbolCard = ({ id, onClick, price, selected, neutral }: SymbolCardProps) 
     (state) => state.stocks.entities[id]
   );
   const showCardInfo = useAppSelector(selectShowCardInfo);
+  const { shake, priceChangeClass } = useCardShake(price);
 
-  const [prevPrice, setPrevPrice] = useState(price);
-  const [shake, setShake] = useState(false);
-  const [priceChangeClass, setPriceChangeClass] = useState('');
+  const cardClass = getCardClass(selected, neutral);
 
   const handleOnClick = () => onClick(id);
-
-  const cardClass = selected
-    ? 'symbolCard--scaleUp'
-    : neutral
-    ? 'symbolCard--neutral'
-    : 'symbolCard--scaleDown';
-
-  useEffect(() => {
-    const priceDiff = price - prevPrice;
-    const percentageChange = (priceDiff / prevPrice) * 100;
-
-    if (priceDiff > 0) {
-      setPriceChangeClass('symbolCard__up');
-    } else if (priceDiff < 0) {
-      setPriceChangeClass('symbolCard__down');
-    } else {
-      setPriceChangeClass('');
-    }
-
-    if (percentageChange > 25) {
-      setShake(true);
-      setTimeout(() => setShake(false), 700);
-    }
-
-    setPrevPrice(price);
-  }, [price]);
 
   return (
     <div
