+++ frontend/src/utils/get-card-class.ts
@@ -1,0 +1,5 @@
+export const getCardClass = (selected: boolean, neutral: boolean) => {
+  if (selected) return 'symbolCard--scaleUp';
+  if (neutral) return 'symbolCard--neutral';
+  return 'symbolCard--scaleDown';
+};
